LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-dialog-element/modal-dialog-blocks-mouse-events-expected.txt

To test manually, move the mouse to the blue box, click, and then move the mouse outside. Then repeat for the red box. The test succeeds if both boxes turn green


FAIL Ensure that mouse events are not dispatched to an inert node. assert_equals: Clicking on inert box expected 5 but got 0
